Refactor to avoid mapping between entities and widget ids
