+++ moxin-frontend/src/app.rs
@@ -2,6 +2,7 @@ use crate::chat::chat_panel::ChatPanelAction;
 use crate::data::store::*;
 use crate::landing::download_item::DownloadItemAction;
 use crate::landing::model_card::{ModelCardViewAllModalWidgetRefExt, ViewAllModalAction};
+use crate::landing::model_files_item::ModelFileItemAction;
 use crate::landing::model_files_items::ModelFileItemsAction;
 use crate::my_models::delete_model_modal::{DeleteModelAction, DeleteModelModalWidgetRefExt};
 use crate::my_models::downloaded_files_table::DownloadedFileAction;
@@ -169,6 +170,7 @@ impl LiveRegister for App {
         // Landing
         crate::landing::shared::live_design(cx);
         crate::landing::model_files_items::live_design(cx);
+        crate::landing::model_files_item::live_design(cx);
         crate::landing::model_files_list::live_design(cx);
         crate::landing::model_files_tags::live_design(cx);
         crate::landing::model_card::live_design(cx);
@@ -243,10 +245,14 @@ impl MatchEvent for App {
             }
 
             match action.as_widget_action().cast() {
-                ModelFileItemsAction::Download(file, model) => {
+                ModelFileItemAction::Download(file, model) => {
                     self.store.download_file(file, model);
                     self.ui.redraw(cx);
                 }
+                _ => {}
+            }
+
+            match action.as_widget_action().cast() {
                 ModelFileItemsAction::Downloaded(file_id) => {
                     let mut popup = self.ui.popup(id!(popup_download_success));
                     popup.set_file_id(file_id);
